+++ src/transaction/Transaction.js
@@ -785,7 +785,7 @@ export default class Transaction extends Executable {
      */
     async freezeWithSigner(signer) {
         await signer.populateTransaction(this);
-        await this.signWithSigner(signer);
+        this.freeze();
         return this;
     }