+++ test/unit/AccountId.js
@@ -37,6 +37,18 @@ describe("AccountId", function () {
         );
     });
 
+    it("clones with alias key", function () {
+        expect(
+            AccountId.fromString(
+                "1.2.302a300506032b657003210008d5a4eebdb9b8451b64d8ad1ff502b493590e513e5e9c9f810dd3258f298542"
+            )
+                .clone()
+                .toString()
+        ).to.be.equal(
+            "1.2.302a300506032b657003210008d5a4eebdb9b8451b64d8ad1ff502b493590e513e5e9c9f810dd3258f298542"
+        );
+    });
+
     it("should construct from (shard, realm, num)", function () {
         const accountId = new AccountId(10, 50, 25050);